Update index.html
This version is better idk if the test version I had was broken or not. This new version saved a randomly generated UUID (Universally Unique Identifier) to server as a session ID to keep users logged in while maintaining security and stuff. Also made usernames case insensitive on the username slot